# perf: update benchmarks