```
     Running bench/parse.rs (target/release/deps/parse-90dae239da4bd51f)
parse/ada_url           time:   [2.0790 µs 2.0812 µs 2.0835 µs]
                        thrpt:  [369.84 MiB/s 370.25 MiB/s 370.65 MiB/s]
                 change:
                        time:   [+0.2255% +0.5063% +0.8385%] (p = 0.00 < 0.05)
                        thrpt:  [-0.8315% -0.5038% -0.2250%]
                        Change within noise threshold.
Found 8 outliers among 100 measurements (8.00%)
  5 (5.00%) high mild
  3 (3.00%) high severe
parse/url               time:   [6.9266 µs 6.9677 µs 7.0199 µs]
                        thrpt:  [109.77 MiB/s 110.59 MiB/s 111.25 MiB/s]
                 change:
                        time:   [-0.3970% +0.0178% +0.5103%] (p = 0.94 > 0.05)
                        thrpt:  [-0.5077% -0.0178% +0.3986%]
                        No change in performance detected.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe

```